+++ complete/src/main/java/com/zhubo/task/parsepage/task/ParseRoomPageWithRoundIdentifyTask.java
@@ -114,8 +114,8 @@ public class ParseRoomPageWithRoundIdentifyTask extends BaseParsePageTask {
                 Long audienceAliasId = StringUtils.isNullOrEmpty(itemElement
                         .getChildText("vipuserid")) ? null : Long.valueOf(itemElement
                         .getChildText("vipuserid"));
-                Integer money = StringUtils.isNullOrEmpty(itemElement.getChildText("top_money")) ? null
-                        : Integer.valueOf(itemElement.getChildText("top_money"));
+                Long money = StringUtils.isNullOrEmpty(itemElement.getChildText("top_money")) ? null
+                        : Long.valueOf(itemElement.getChildText("top_money"));
                 if (audienceName != null && audienceAliasId != null
                         && !pays.containsKey(audienceAliasId)) {
                     pays.put(audienceAliasId, new Pay(audienceAliasId, audienceName, money));
@@ -186,7 +186,7 @@ public class ParseRoomPageWithRoundIdentifyTask extends BaseParsePageTask {
     }
 
     private void storePayPeriodAndPayMinute(ResourceManager rm, long audienceId, long anchorId,
-            int platformId, boolean isOldRound, Date latestRoundStart, int roundIncome, int periodMoney, Date ts) {
+            int platformId, boolean isOldRound, Date latestRoundStart, int roundIncome, long periodMoney, Date ts) {
         Date periodStart = getQixiuPayAggregateDate(ts);
         PayPeriodObject payPeriod = new PayPeriodObject(platformId, periodMoney, periodStart, ts);
         Integer diffMoney = resourceManager.getDatabaseCache()
@@ -219,7 +219,7 @@ public class ParseRoomPageWithRoundIdentifyTask extends BaseParsePageTask {
     }
 
     private void storePayPeriodIfNeeded(ResourceManager rm, long audienceId, long anchorId,
-            int platformId, int money, Date ts, Date periodStart) {
+            int platformId, long money, Date ts, Date periodStart) {
         if (!rm.getDatabaseCache().existInPayPeriod(audienceId, anchorId, ts)) {
             AudiencePayPeriod payPeriod = new AudiencePayPeriod(audienceId, anchorId, platformId,
                     money, ts, periodStart);
@@ -325,9 +325,9 @@ public class ParseRoomPageWithRoundIdentifyTask extends BaseParsePageTask {
     public static class Pay {
         public Long audienceAliasId;
         public String audienceName;
-        public Integer money;
+        public Long money;
 
-        public Pay(Long audienceAliasId, String audienceName, Integer money) {
+        public Pay(Long audienceAliasId, String audienceName, Long money) {
             this.audienceAliasId = audienceAliasId;
             this.audienceName = audienceName;
             this.money = money;
